Move get_missing_keyboard_configuration method
This method is not directly bound to localed and will be enhanced to
also take into account Live system configuration. So let's move it first.
